Handle embedded references in samtools index & idxstats
Adds a third entry to the BAM_MAX_BIN array in BAI files.
Entry zero stores the start/end offsets for the placed but
unmapped reads. Entry one stores the mapped and unmapped
(but placed) read counts. New entry two stores the start/
end offsets of any embedded reference sequence.

Note that older versions of samtools should happily
ignore this new entry in the BAI file.

Updated 'samtools idxstats' to include another column,
the number of reads representing the embedded reference.
This will be 0 or 1, or in the case of an out of date
BAI file, a question mark is shown.

TODO: Use this new index information to provide an
alternative to FAI indexed FASTA files for accessing
the reference sequence.